Refactored code (Program Menu). Made the menu validated and cleaner.

+++ libs/class_particleInitHelper/particleInitHelper.h
@@ -2,12 +2,11 @@
 #define PARTICLE_INIT_HELPER
 
 #include "glm/ext/vector_float3.hpp"
-#include <cstddef>
 #include <string>
 
 class class_particleInitHelper{
 public:
-    class_particleInitHelper(size_t& count_particleTypes);
+    class_particleInitHelper(int& count_particleTypes);
     ~class_particleInitHelper();
 
     std::string* name;
